refactor: extract shared coerceToNumber utility to eliminate duplication
The fold projection's coerceToNumber and span mapper's asNumberOrNull
were functionally identical. Extract to ~/utils/coerceToNumber.ts and
update both consumers to import from the shared location.
